# Add inference-time sequence packing support