--- a/gliner/__init__.py
+++ b/gliner/__init__.py
@@ -1,9 +1,21 @@
 __version__ = "0.2.22"
 
 from .model import GLiNER
+from .infer_packing import (
+    InferencePackingConfig,
+    PackedBatch,
+    pack_requests,
+    unpack_spans,
+)
 from .config import GLiNERConfig
 # from .multitask import (GLiNERClassifier, GLiNERQuestionAnswerer, GLiNEROpenExtractor,
 #                                 GLiNERRelationExtractor, GLiNERSummarizer, GLiNERSquadEvaluator,
 #                                     GLiNERDocREDEvaluator)
 
-__all__ = ["GLiNER"]
+__all__ = [
+    "GLiNER",
+    "InferencePackingConfig",
+    "PackedBatch",
+    "pack_requests",
+    "unpack_spans",
+]