## Summary
- Add a new **inference-packing utility** that builds packed batches with block-diagonal masks and helpers to unpack outputs.  
- Expose configuration knobs on the GLiNER API so packing can be toggled globally or per-call.  
- Wire the encoder to use packed execution when configured, including automatic pack/unpack around the transformer forward pass.  
- Results are **identical with or without packing** (verified by tests).  

## Benchmarks  
All runs on **CPU**, `roberta-base`, `batch_size=64`, `max_length=512`.  

| Scenario       | Baseline tokens/s | Packed tokens/s | Speedup | Padding ↓       |
|----------------|------------------:|----------------:|--------:|----------------:|
| short_zipf     | 2.00e+03          | 3.88e+03        | **1.94×** | 61.5% → 12.0%  |
| short_uniform  | 2.44e+03          | 3.47e+03        | **1.42×** | 45.9% → 13.4%  |
| mixed_tail     | 6.18e+02          | 3.43e+03        | **5.55×** | 87.4% → 19.6%  |
| flat_long      | 4.94e+03          | 4.10e+03        | 0.83×   | 0.0% → 0.0%    |

👉 Packing yields **1.4–5.5× throughput improvements** when input lengths are short or skewed, while performance is neutral (or slightly worse) when all sequences are long and uniform.